fix: Users can only delete or update their own comments

+++ src/server-actions/comments/delete-comment.ts
@@ -4,18 +4,18 @@ import { db } from '@server/database';
 import { comments } from '@server/database/schemas/comments';
 import { getCurrentUser } from '@server/helpers/get-current-user';
 import type { ServerActionResponse } from '@server-actions/common/server-action-response';
-import { eq } from 'drizzle-orm';
+import { and, eq } from 'drizzle-orm';
 import { getExtracted } from 'next-intl/server';
 
 export const deleteComment = async (commentId: number): Promise<ServerActionResponse> => {
     const t = await getExtracted('common');
     try {
-        const currentUser = await getCurrentUser();
-        if (!currentUser) {
+        const user = await getCurrentUser();
+        if (!user) {
             throw new Error('Unauthorized');
         }
 
-        await db.delete(comments).where(eq(comments.id, commentId));
+        await db.delete(comments).where(and(eq(comments.id, commentId), eq(comments.userId, user.id)));
         return {};
     } catch (e) {
         console.error(e);
